requirements.txt
PyQt5
asyncqt
PyYAML
aiohttp